Move parseAt to new unit test
